+++ src/main/scala/expressive/parser/Token.scala
@@ -14,7 +14,7 @@ case object Open extends Evaluable with Operator {
 
 case object Close extends Evaluable
 
-case class Number(value: Int) extends Evaluable
+case class Number(value: Double) extends Evaluable
 
 case object Divide extends Evaluable with Operator {
   override val precedence: Int = 3
